requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
astroid==2.0.4
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
DateTime==4.3
decorator==4.3.0
dominate==2.3.4
emoji==0.5.1
feedfinder2==0.0.4
feedparser==5.2.1
flake8==3.5.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Session==0.3.1
google==2.0.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gspread==3.0.1
httplib2==0.11.3
idna==2.7
isort==4.3.4
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.3
numpy==1.15.3
oauth2client==4.1.3
oauthlib==2.1.0
pbr==5.0.0
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pythonwhois==2.4.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
requests-oauthlib==1.0.0
rsa==4.0
selenium==3.14.1
six==1.11.0
SQLAlchemy==1.2.12
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
stemming==1.0
Tempita==0.5.2
tinysegmenter==0.3
tldextract==2.2.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
visitor==0.1.3
webapp2==2.5.2
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
youtube-python==1.0.12
youtube-transcript-api==0.1.1
zope.interface==4.5.0
flask-dance==1.1.0
Flask-SSLify==0.1.5
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1